using and implementing more interfaces from `connector` in `apns` ; extracting apns_sender.go; moving queue to `connector` package
